LPS-138586 Not all images use company 0
Some images are already put under the correct company; skip these to
avoid upgrade failures.

+++ modules/apps/image/image-impl/src/main/java/com/liferay/image/internal/upgrade/v1_0_0/ImageStorageUpgradeProcess.java
@@ -51,20 +51,25 @@ public class ImageStorageUpgradeProcess extends UpgradeProcess {
 			(Image image) -> {
 				String fileName = _getFileName(image);
 
-				try (InputStream inputStream = store.getFileAsStream(
+				if (store.hasFile(
 						CompanyConstants.SYSTEM, _REPOSITORY_ID, fileName,
 						StringPool.BLANK)) {
 
-					store.addFile(
-						image.getCompanyId(), _REPOSITORY_ID, fileName,
-						Store.VERSION_DEFAULT, inputStream);
+					try (InputStream inputStream = store.getFileAsStream(
+							CompanyConstants.SYSTEM, _REPOSITORY_ID, fileName,
+							StringPool.BLANK)) {
 
-					store.deleteFile(
-						CompanyConstants.SYSTEM, _REPOSITORY_ID, fileName,
-						Store.VERSION_DEFAULT);
-				}
-				catch (Exception exception) {
-					_log.error(exception, exception);
+						store.addFile(
+							image.getCompanyId(), _REPOSITORY_ID, fileName,
+							Store.VERSION_DEFAULT, inputStream);
+
+						store.deleteFile(
+							CompanyConstants.SYSTEM, _REPOSITORY_ID, fileName,
+							Store.VERSION_DEFAULT);
+					}
+					catch (Exception exception) {
+						_log.error(exception, exception);
+					}
 				}
 			});
